requirements.txt
torch
torchvision
captum
matplotlib
opencv-python
numpy
Pillow
streamlit
scipy
tqdm
pathlib
pandas
colorama